requirements.txt
# Core LangChain + LLM + Embeddings
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
tiktoken

# Vector store (FAISS)
faiss-cpu>=1.7.4
langchain-community>=0.2.0

# Reranking model
sentence-transformers>=2.2.2
transformers>=4.41.0

# Optional but useful
python-dotenv
numpy
scikit-learn

